+++ pubspec.yaml
@@ -24,7 +24,6 @@ dependencies:
   url_launcher: ^6.3.1
   file_picker: ^8.1.7
   qr_flutter: ^4.1.0
-  mobile_scanner: ^6.0.7
   shared_preferences: ^2.3.5
   collection: ^1.19.1
   ffi: ^2.1.3
@@ -37,6 +36,8 @@ dependencies:
   super_sliver_list: ^0.4.1
   defer_pointer: ^0.0.2
   downloadsfolder: ^1.2.0
+  open_filex: ^4.7.0
+  gscankit: ^1.0.3
 
 dev_dependencies:
   flutter_test:
